==> 0_gensim_entity_embedding.ipynb <==
"""
todo
三种方法：
1 gensim 训练　embedding
2 openKE 训练　embedding
3 openKE　分词训练　embedding　
"""

# coding:u8
import json
import numpy as np
from collections import defaultdict
from tqdm import tqdm as tqdm
from gensim.models import word2vec
import jieba

# 读取知识库
kb_data = []
kb = {}
with open('data/raw_data/kb_data','r') as f:
    for line in f:
        item = eval(str(json.loads(line)).lower())
        kb[item['subject_id']] = item
        kb_data.append(item)


# OpenKE 默认方法

predicate = {}
subject = {}
type_object = {}

# for item in tqdm(kb_data):
#     if item['subject_id'] not in subject:
#         subject[item['subject_id']] = len(subject)
#     for t in item['data']:
#         if t['predicate'] == '摘要':continue
#         if t['predicate'] not in predicate:
#             predicate[t['predicate']] = len(predicate)
#         if t['object'] not in type_object:
#             type_object[t['object']] = len(type_object)
#     for t_type in item['type']:
#         if t_type not in type_object:
#             type_object[t_type] = len(type_object)
# if 'type' not in predicate:
#     predicate['type'] = len(predicate)

# spo = []
# for item in tqdm(kb_data):
#     subject_id = subject[item['subject_id']]
#     for t_type in item['type']:
#         spo.append([subject_id,predicate['type'],type_object[t_type]])
#     for p in item['data']:
#         if p['predicate'] == '摘要':continue
#         spo.append([subject_id,predicate[p['predicate']],type_object[p['object']]])

# with open('entity_embedding/train2id.txt','w') as f:
#     #0 1 0
#     f.write(str(len(spo))+'\n')
#     for spo_line in spo:
#         f.write(str(spo_line[0])+' '+str(spo_line[1])+' '+str(spo_line[2])+'\n')

# # 14951
# # /m/027rn	0
# with open('entity_embedding/entity2id.txt','w') as f:
#     f.write(str(len(type_object))+'\n')
#     for key in type_object:
#         f.write(str(key)+' '+str(type_object[key])+'\n')

# with open('entity_embedding/relation2id.txt','w') as f:
#     f.write(str(len(predicate))+'\n')
#     for key in predicate:
#         f.write(str(key)+' '+str(predicate[key])+'\n')

# gensim method

# jieba加字典
for item in tqdm(kb_data):
    jieba.add_word(item['subject'],freq=10000)
    for name in item['alias']:
        jieba.add_word(name, freq=10000)



sentence = []
for item in tqdm(kb_data):
    sentence.append([item['subject_id'],item['type'][0]])
    for t in item['data']:
        temp = []
        if t['predicate'] == '摘要':
            abstract = jieba.lcut(t['object'])
            for word in abstract:
                if word !=item['subject']:
                    temp.append(word)
                else:
                    temp.append(item['subject_id'])
            sentence.append(temp)
        else:
            temp.append(item['subject_id'])
            temp.append(t['predicate'])
            for w in jieba.lcut(t['object']):
                temp.append(w)
            sentence.append(temp)

train_data = []
with open('data/raw_data/train.json','r') as f:
    for line in f:
        sentence.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))
with open('data/raw_data/develop.json','r') as f:
    for line in f:
        sentence.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))
with open('data/raw_data/eval722.json','r') as f:
    for line in f:
        sentence.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))

model = word2vec.Word2Vec(sentence, size=100, window=5, min_count=1, workers=10)
# 打印单词'good'的词向量
print(model.wv.word_vec('good'))
# 打印和'good'相似的前2个单词
print(model.wv.most_similar('good', topn=2))
# 保存模型到文件
model.save('embedding/w2v.model')

from gensim.models import Word2Vec

entity_embedding = Word2Vec.load('embedding/w2v.model')
print(entity_embedding.wv.word_vec('搜索'))
