==> 1_feature_engineering.ipynb <==
# coding:u8
import json
import numpy as np
import pandas as pd
from tqdm import tqdm as tqdm
import gc,jieba
from gensim.models import Word2Vec
from sklearn.model_selection import KFold

# train_data = []
# train_data_dict = {}
# with open('data/raw_data/train.json','r') as f:
#     for line in f:
#         raw = json.loads(line)
#         train_data.append(raw)
#         train_data_dict[raw['text_id']] = raw
# # 读取知识库
# kb_data = []
# kb = {}
# with open('data/raw_data/kb_data','r') as f:
#     for line in f:
#         item = json.loads(line)
#         kb[item['subject_id']] = item
#         kb_data.append(item)


train_data = []
train_data_dict = {}
with open('data/raw_data/train.json', 'r') as f:
    for line in f:
        raw = eval(str(json.loads(line)).lower())
        train_data.append(raw)
        #train_data_dict[raw['text_id']] = raw
kb = {}
with open('data/raw_data/kb_data', 'r') as f:
    for line in f:
        item = eval(str(json.loads(line)).lower())
        kb[item['subject_id']] = item

name_id = {}
for kb_id in kb:
    for item in kb[kb_id]['alias']:
        if item not in name_id:
            name_id[item] = [kb_id]
        else:
            name_id[item].append(kb_id)
    if kb[kb_id]['subject'] not in name_id:
        name_id[kb[kb_id]['subject']] = [kb_id]
    else:
        name_id[kb[kb_id]['subject']].append(kb_id)
for id in name_id:
    name_id[id] = sorted(list(set(name_id[id])))

"""
TODO:
预测type,
引入tfidf
"""

def get_common_word(a, b):
        common_count = 0
        for word in a:
            if word in b:
                common_count += 1
        return common_count
def cal_similarity(params):
    index,sentence_a,sentence_b = params
    la = len(sentence_a)
    lb = len(sentence_b)

    char_a = [c for c in sentence_a]
    char_b = [c for c in sentence_b]
    word_a = jieba.lcut_for_search(sentence_a)
    word_b = jieba.lcut_for_search(sentence_b)
    bigram_a = [sentence_a[i:i + 2] for i in range(len(sentence_a) - 1)]
    bigram_b = [sentence_b[i:i + 2] for i in range(len(sentence_b) - 1)]
    trigram_a = [sentence_a[i:i + 3] for i in range(len(sentence_a) - 2)]
    trigram_b = [sentence_b[i:i + 3] for i in range(len(sentence_b) - 2)]

    common_word, common_word_score = 0, 0
    for w in word_b:
        if w in word_a:
            common_word += 1
            common_word_score += 1

    common_c = get_common_word(char_a, char_b)
    common_bigram = get_common_word(bigram_a, bigram_b)
    common_trigram = get_common_word(trigram_a, trigram_b)

    max_len = max(la, lb)
    min_len = min(la, lb)
    if max_len==0:
        max_len=0.01
    if min_len==0:
        min_len = 0.01
    numerical_feats = [max_len, min_len, common_word, common_word_score,
                                common_word / min_len, common_word / max_len,
                                common_word_score / min_len, common_word_score / max_len,
                                common_c / min_len, common_c / max_len,
                                common_bigram / min_len, common_bigram / max_len,
                                common_trigram / min_len, common_trigram / max_len,
                                common_word_score - common_word]
    return index, numerical_feats

def is_works(mention,text):
    if '《'+mention in text or mention+'》' in text:
        return 1
    elif '【'+mention+'】' in text:
        return 1
    elif '['+mention+']' in text:
        return 1
    else:
        return 0

# 书名号纠错

# 提取candidate候选特征
kb_column = []
train_column = []
text_id = []
label_colum = []
num_attrs = []
num_abstract_words = []
num_alias = []
m_id = 0
m_id_list = []
num_candidate = []
equal_subject = []
shuminghao = [] 
entity_common = []
len_mention = []
mention_start = []
mention_end = []
numerical_f = []
is_eng = []
create_works = []
end_with_point = []
end_with_line = []
end_with_line2 = []
end_with_shuminghao = []
for s in tqdm(train_data):
    mention_ner = s['mention_data']
    m_list = []
    for m in mention_ner:
        m_list.append(m['mention'])
    for m in mention_ner:
        
        if m['mention'] not in name_id : # 不能注销，因为有些实体没有or m['kb_id']=='nil':
            continue
        if m['kb_id'] != 'nil':
            name_list = [kb[m['kb_id']]['subject']]+kb[m['kb_id']]['alias']
            if m['mention'] not in name_list:
                    continue

        candidate_ids = name_id[m['mention']]
        for m_candidate_id in candidate_ids:
            # 统计配对特征
            candidate_numattrs = 0
            candidate_abstract_numwords = 0
            candidate_detail = kb[m_candidate_id]
            candi_text = ''
            for predicate in candidate_detail['data']:
                candidate_numattrs += 1
                candi_text += predicate['object']
                #candi_text += predicate['predicate']

                if predicate['predicate'] == '摘要':
                    candidate_abstract = predicate['object']
                    candidate_abstract_numwords = len(candidate_abstract)
#                 if predicate['predicate'] == '标签':
#                     candidate_label += predicate['object']
            if m_candidate_id==m['kb_id']:
                label = 1
            else:
                label = 0
            train_column.append(m['mention'])
            kb_column.append(m_candidate_id)
            text_id.append(s['text_id'])
            label_colum.append(label)
            num_attrs.append(candidate_numattrs)
            num_abstract_words.append(candidate_abstract_numwords)
            candidate_name_list = candidate_detail['alias']+[candidate_detail['subject']]
            num_alias.append(len(set(candidate_name_list)))
            m_id_list.append(m_id)
            num_candidate.append(len(candidate_ids))
            equal_subject.append(candidate_detail['subject']==m['mention'])
            
            shu_in_abstract = False
            for sub_alia in candidate_name_list:
                if '《'+sub_alia+'》' in candidate_abstract:
                    shu_in_abstract = True
            shuminghao.append(shu_in_abstract)
            
#             if langid.classify(m['mention'])[0]=='en':
#                 len_mention.append(len(m['mention'].split()))
#                 is_eng.append(1)
#             else:
#                 len_mention.append(len(m['mention']))
#                 is_eng.append(0)
            len_mention.append(len(m['mention']))
            mention_start.append(m['offset'])
            mention_end.append(len(s['text'])-int(m['offset'])-len(m['mention']))
            
            
            i = 0
            for item in m_list:
                if item in candi_text: i+=1
            entity_common.append(i)
            create_works.append(is_works(m['mention'],s['text']))
            end_with_point.append('.' == s['text'][-1])
            end_with_line.append('-' in s['text'])
            end_with_line2.append('_' in s['text'])
            end_with_shuminghao.append('《' in s['text'])
        m_id += 1
        
data = pd.DataFrame()
data['text_id'] = text_id
data['kb_id'] = kb_column
data['train_mention'] = train_column
data['label'] = label_colum
data['num_attrs'] = num_attrs
data['num_abstract_words'] = num_abstract_words
data['num_alias'] = num_alias
data['m_id'] = m_id_list
data['num_candidates'] = num_candidate 
data['mention_equal_subject'] = equal_subject
data['shuminghao'] = shuminghao
data['entity_common']  = entity_common
data['len_mention'] = len_mention
data['mention_start'] = list(map(int,mention_start))
data['create_works'] = create_works
data['mention_end'] = mention_end
data['end_with_point'] = end_with_point
data['end_with_line'] = end_with_line
data['end_with_line2'] = end_with_line2
data['end_with_shuminghao'] = end_with_shuminghao
data['works_match'] = data.apply(lambda x: 1 if x.create_works==x.shuminghao else 0,axis=1)

1==False

print(data.shape)


# 提取candidate候选特征
numerical_f = []
i = 0
for s in tqdm(train_data):
    mention_ner = s['mention_data']
    m_list = []
    for m in mention_ner:
        m_list.append(m['mention'])
    for m in mention_ner:
        if m['mention'] not in name_id : # 不能注销，因为有些实体没有or m['kb_id']=='nil':
            continue
        if m['kb_id'] != 'nil':
            name_list = [kb[m['kb_id']]['subject']]+kb[m['kb_id']]['alias']
            if m['mention'] not in name_list:
                    continue
        
        candidate_ids = name_id[m['mention']]
#         add_extra = 0 
#         if m['kb_id'] != 'nil':
#             name_list = [kb[m['kb_id']]['subject']]+kb[m['kb_id']]['alias']
            
#             if m['mention'] not in name_list:
#                 keep = False
#                 for name_c in ''.join(name_list):
#                     if name_c in m['mention']:
#                         #print(name_c,m['mention'])
#                         keep = True
#                         break
#                 if not keep:
#                     continue
#                 candidate_ids = name_id[m['mention']]+[m['kb_id']]
#                 add_extra = 1
        for m_candidate_id in candidate_ids:
            # 统计配对特征
            candidate_numattrs = 0
            candidate_abstract_numwords = 0
            candidate_detail = kb[m_candidate_id]
            candi_text = ''
            for predicate in candidate_detail['data']:
                candidate_numattrs += 1
                candi_text += predicate['object']
                #candi_text += predicate['predicate']
            numerical_f.append((i,candi_text,s['text']))
            i+=1
            #numerical_f.append(cal_similarity(candi_text,s['text']))


import multiprocessing
p = multiprocessing.Pool(6)
res = p.map(cal_similarity, numerical_f)
res = sorted(res,key=lambda x:x[0],reverse=False)

nume_f = [x[1] for x in res]
numerical_f = np.array(nume_f)
for i in range(numerical_f.shape[1]):
    data['numerical_%d'%i] = numerical_f[:,i]

data.head()

# f_columns = ['numerical_%d'%i for i in range(15)]

# data[f_columns].to_pickle('data/numerical_f.pkl')

# candidate　type的one-hot特征
data['type'] = data.apply(lambda x:kb[x['kb_id']]['type'][0],axis=1)
#data = pd.concat([data,pd.get_dummies(data['type'])],axis=1)

# target encoding
groupbytype = data.groupby('type',as_index=False)['label'].agg({
    'type_label_mean':'mean','type_label_count':'count'})
data = data.merge(groupbytype,on='type',how='left')

data.to_pickle('data/step1.pkl')

max_mid = data['m_id'].max()

# 载入测试数据

# # 戴安楠ner
# ner_result = pd.read_csv('result/el_dev_result(1).csv',sep='\t')
# ner_result['offset'] = ner_result['offset'].apply(lambda x:eval(x.lower()))
# ner_result['mention'] = ner_result['mention'].apply(lambda x:eval(x.lower()))
# ner_result = ner_result.rename(columns={'mention':'pred','offset':'pos'})


# my ner
ner_result = pd.read_pickle('result/ner_bert_result.pkl')
ner_result['text'] = ner_result['text'].apply(lambda x:''.join(x[1:-1]))
ner_result['pos'] = ner_result['pos'].apply(lambda x:[k[0]-1 for k in x])
ner_result['text_id'] = ner_result.index+1

ner_result.head()

# 包装成和train_data一样的格式
test_data = []
for index,row in ner_result.iterrows():
    item = {}
    item['text_id'] = row['text_id']
    item['text'] = row['text']
    item['mention_data'] = []
    for i,p in enumerate(row['pos']):
        item['mention_data'].append({'mention':row['pred'][i], 'offset':str(p)})
    test_data.append(item)

# 提取candidate候选特征
kb_column = []
train_column = []
text_id = []
num_attrs = []
num_abstract_words = []
num_alias = []
m_id = 0
m_id_list = []
num_candidate = []
equal_subject = []
shuminghao = [] 
entity_common = []
len_mention = []
mention_start = []
mention_end = []
numerical_f = []
is_eng = []
create_works = []
end_with_point = []
end_with_line = []
end_with_line2 = []
end_with_shuminghao = []
for s in tqdm(test_data):
    mention_ner = s['mention_data']
    m_list = []
    for m in mention_ner:
        m_list.append(m['mention'])
    
    for m in mention_ner:    
        if m['mention'] not in name_id : # 不能注销，因为有些实体没有or m['kb_id']=='nil':
            continue
        candidate_ids = name_id[m['mention']]
        for m_candidate_id in candidate_ids:
            # 统计配对特征
            candidate_numattrs = 0
            candidate_abstract_numwords = 0
            candidate_detail = kb[m_candidate_id]
            candi_text = ''
            for predicate in candidate_detail['data']:
                candidate_numattrs += 1
                candi_text += predicate['object']
                #candi_text += predicate['predicate']
                if predicate['predicate'] == '摘要':
                    candidate_abstract = predicate['object']
                    candidate_abstract_numwords = len(candidate_abstract)
#                 if predicate['predicate'] == '标签':
#                     candidate_label += predicate['object']
            train_column.append(m['mention'])
            kb_column.append(m_candidate_id)
            num_attrs.append(candidate_numattrs)
            num_abstract_words.append(candidate_abstract_numwords)
            candidate_name_list = candidate_detail['alias']+[candidate_detail['subject']]
            num_alias.append(len(set(candidate_name_list)))
            m_id_list.append(m_id)
            num_candidate.append(len(candidate_ids))
            equal_subject.append(candidate_detail['subject']==m['mention'])
            text_id.append(s['text_id'])
            shu_in_abstract = False
            for sub_alia in candidate_name_list:
                if '《'+sub_alia+'》' in candidate_abstract:
                    shu_in_abstract = True
            shuminghao.append(shu_in_abstract)
#             if langid.classify(m['mention'])[0]=='en':
#                 len_mention.append(len(m['mention'].split()))
#                 is_eng.append(1)
#             else:
#                 len_mention.append(len(m['mention']))
#                 is_eng.append(0)
            len_mention.append(len(m['mention']))
            mention_start.append(m['offset'])
            mention_end.append(len(s['text'])-int(m['offset'])-len(m['mention']))
            
            
            i = 0
            for item in m_list:
                if item in candi_text: i+=1
            entity_common.append(i)
            create_works.append(is_works(m['mention'],s['text']))
            end_with_point.append('.' == s['text'][-1])
            end_with_line.append('-' in s['text'])
            end_with_line2.append('_' in s['text'])
            end_with_shuminghao.append('《' in s['text'])
        m_id += 1
        
data = pd.DataFrame()
data['text_id'] = text_id
data['kb_id'] = kb_column
data['train_mention'] = train_column
data['num_attrs'] = num_attrs
data['num_abstract_words'] = num_abstract_words
data['num_alias'] = num_alias
data['m_id'] = m_id_list
data['num_candidates'] = num_candidate 
data['mention_equal_subject'] = equal_subject
data['shuminghao'] = shuminghao
data['entity_common']  = entity_common
data['len_mention'] = len_mention
data['mention_start'] = list(map(int,mention_start))
data['create_works'] = create_works
data['mention_end'] = mention_end
data['end_with_point'] = end_with_point
data['end_with_line'] = end_with_line
data['end_with_line2'] = end_with_line2
data['end_with_shuminghao'] = end_with_shuminghao
data['works_match'] = data.apply(lambda x: 1 if x.create_works==x.shuminghao else 0,axis=1)

# 提取candidate候选特征
numerical_f = []
i = 0
for s in tqdm(test_data):
    mention_ner = s['mention_data']
    m_list = []
    for m in mention_ner:
        m_list.append(m['mention'])
    for m in mention_ner:
        if m['mention'] not in name_id : # 不能注销，因为有些实体没有or m['kb_id']=='nil':
            continue
        candidate_ids = name_id[m['mention']]

        for m_candidate_id in candidate_ids:
            # 统计配对特征
            candidate_numattrs = 0
            candidate_abstract_numwords = 0
            candidate_detail = kb[m_candidate_id]
            candi_text = ''
            for predicate in candidate_detail['data']:
                candidate_numattrs += 1
                candi_text += predicate['object']
                #candi_text += predicate['predicate']
            numerical_f.append((i,candi_text,s['text']))
            i+=1
            #numerical_f.append(cal_similarity(candi_text,s['text']))


import multiprocessing
p = multiprocessing.Pool(6)
res = p.map(cal_similarity, numerical_f)
res = sorted(res,key=lambda x:x[0],reverse=False)

nume_f = [x[1] for x in res]
numerical_f = np.array(nume_f)
for i in range(numerical_f.shape[1]):
    data['numerical_%d'%i] = numerical_f[:,i]

# candidate　type的one-hot特征
data['type'] = data.apply(lambda x:kb[x['kb_id']]['type'][0],axis=1)
#data = pd.concat([data,pd.get_dummies(data['type'])],axis=1)

# target encoding
data = data.merge(groupbytype,on='type',how='left')

data['m_id'] = data['m_id']+max_mid+1

data.reset_index(drop=True,inplace=True)
data.to_pickle('data/step1_test.pkl')

data

import pandas as pd
test = pd.read_pickle('data/step1_test.pkl')

test['mention_start']
