==> 1.5_feature_engineering.ipynb <==
# 数值特征
import pandas as pd
import numpy as np
import json
from tqdm import tqdm as tqdm
from sklearn.model_selection import KFold

train_data = []
train_data_dict = {}
with open('data/raw_data/train.json', 'r') as f:
    for line in f:
        raw = eval(str(json.loads(line)).lower())
        train_data.append(raw)
        #train_data_dict[raw['text_id']] = raw
kb = {}
with open('data/raw_data/kb_data', 'r') as f:
    for line in f:
        item = eval(str(json.loads(line)).lower())
        kb[item['subject_id']] = item

data = pd.read_pickle('data/step1.pkl')
test_data = pd.read_pickle('data/step1_test.pkl')

print(data.shape)
print(test_data.shape)


test_data['label'] = -1

data_all = data.append(test_data)

data_all.shape

data.columns

numerical_clolumns = ['num_candidates',
       'mention_equal_subject', 'shuminghao', 'entity_common', 'len_mention',
       'mention_start', 'create_works', 'mention_end', 'end_with_point',
       'end_with_line', 'end_with_line2', 'end_with_shuminghao', 'numerical_0',
       'numerical_1', 'numerical_2', 'numerical_3', 'numerical_4',
       'numerical_5', 'numerical_6', 'numerical_7', 'numerical_8',
       'numerical_9', 'numerical_10', 'numerical_11', 'numerical_12',
       'numerical_13', 'numerical_14', 'type_label_mean',
       'type_label_count','works_match']

# rank 特征
to_rank = numerical_clolumns
for f in to_rank:
    data_group = data_all.groupby('kb_id')[f].rank(ascending=False)
    data_all['kbid_rank_'+f] = data_group

# mean and var
to_rank = numerical_clolumns
for f in tqdm(to_rank):
    data_group = data_all.groupby('kb_id')[f].agg({'kb_'+f+'_var':'var','kb_'+f+'_mean':'mean','kb_'+f+'_max':'max',
                                                   'kb_'+f+'_min':'min'})
    data_all = data_all.merge(data_group,on='kb_id',how='left')

# mean and var
to_rank = numerical_clolumns
for f in tqdm(to_rank):
    data_group = data_all.groupby(['kb_id','train_mention'])[f].agg({'kb_m_'+f+'_var':'var','kb_m_'+f+'_mean':'mean',
                                                                 'kb_m_'+f+'_max':'max',
                                                   'kb_m_'+f+'_min':'min'})
    data_all = data_all.merge(data_group,on=['kb_id','train_mention'],how='left')

# rank 特征
to_rank = numerical_clolumns
for f in to_rank:
    data_group = data_all.groupby(['kb_id','train_mention'])[f].rank(ascending=False)
    data_all['kb_m_rank_'+f] = data_group
   

for item in data_all.columns:
    if data_all.dtypes[item]=='float64':
        print(item)
        data_all[item] = data_all[item].astype('float32')
    if data_all.dtypes[item]=='int64':
        print(item)
        data_all[item] = data_all[item].astype('int32')
    if data_all.dtypes[item]=='bool':
        print(item)
        data_all[item] = data_all[item].astype('int8')

# rank 特征
to_rank = ['num_attrs',
       'num_abstract_words', 'num_alias', 'num_candidates',
       'mention_equal_subject', 'shuminghao', 'entity_common', 'len_mention',
       'mention_start', 'create_works', 'mention_end', 'end_with_point',
       'end_with_line', 'end_with_line2', 'end_with_shuminghao', 'numerical_0',
       'numerical_1', 'numerical_2', 'numerical_3', 'numerical_4',
       'numerical_5', 'numerical_6', 'numerical_7', 'numerical_8',
       'numerical_9', 'numerical_10', 'numerical_11', 'numerical_12',
       'numerical_13', 'numerical_14', 'type_label_mean',
       'type_label_count','works_match']
for f in to_rank:
    data_group = data_all.groupby('m_id')[f].rank(ascending=False)
    data_all['m_rank_'+f] = data_group

data_all.head()

# mean and var
to_rank = ['num_attrs',
       'num_abstract_words', 'num_alias', 'num_candidates',
       'mention_equal_subject', 'shuminghao', 'entity_common', 'len_mention',
       'mention_start', 'create_works', 'mention_end', 'end_with_point',
       'end_with_line', 'end_with_line2', 'end_with_shuminghao', 'numerical_0',
       'numerical_1', 'numerical_2', 'numerical_3', 'numerical_4',
       'numerical_5', 'numerical_6', 'numerical_7', 'numerical_8',
       'numerical_9', 'numerical_10', 'numerical_11', 'numerical_12',
       'numerical_13', 'numerical_14', 'type_label_mean',
       'type_label_count','works_match']
for f in tqdm(to_rank):
    data_group = data_all.groupby('m_id')[f].agg({'mid_'+f+'_var':'var','mid_'+f+'_mean':'mean',
                                                  'mid_'+f+'_max':'max','mid_'+f+'_min':'min'})
    data_all = data_all.merge(data_group,on='m_id',how='left')

# 二阶统计特征 暂定

# data_all.dtypes

for item in data_all.columns:
    if data_all.dtypes[item]=='float64':
        print(item)
        data_all[item] = data_all[item].astype('float32')
    if data_all.dtypes[item]=='int64':
        data_all[item] = data_all[item].astype('int32')
    if data_all.dtypes[item]=='bool':
        data_all[item] = data_all[item].astype('int8')

data = data_all.iloc[0:data.shape[0],:]
test_data = data_all.iloc[data.shape[0]:,:]
data.reset_index(drop=True,inplace=True)
test_data.reset_index(drop=True,inplace=True)

test_data.shape

data.to_pickle('data/step1_5.pkl')

test_data.to_pickle('data/step1_5_test.pkl')
