==> 3_tf_idf_feature.ipynb <==
# 数值特征
import pandas as pd
import numpy as np
import json
from tqdm import tqdm as tqdm
from sklearn.model_selection import KFold
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.decomposition import TruncatedSVD
import jieba
# 获取mention的tfidf
from gensim.models import TfidfModel
from gensim.corpora import Dictionary
import gensim.downloader as api
from gensim.models import LsiModel

train_data = []
train_data_dict = {}
with open('data/raw_data/train.json', 'r') as f:
    for line in f:
        raw = eval(str(json.loads(line)).lower())
        train_data.append(raw)
        train_data_dict[raw['text_id']] = raw
# 读取知识库
kb_data = []
kb = {}
with open('data/raw_data/kb_data','r') as f:
    for line in f:
        item = eval(str(json.loads(line)).lower())
        kb[item['subject_id']] = item
        kb_data.append(item)


name_id = {}
for kb_id in kb:
    for item in kb[kb_id]['alias']:
        if item not in name_id:
            name_id[item] = [kb_id]
        else:
            name_id[item].append(kb_id)
    if kb[kb_id]['subject'] not in name_id:
        name_id[kb[kb_id]['subject']] = [kb_id]
    else:
        name_id[kb[kb_id]['subject']].append(kb_id)
for id in name_id:
    name_id[id] = sorted(list(set(name_id[id])))
for id in name_id:
    jieba.add_word(id,freq=10000)

sentence = []
for item in tqdm(kb_data):
    sentence.append([item['subject_id'],item['type'][0]])
    for t in item['data']:
        temp = []
        if t['predicate'] == '摘要':
            abstract = jieba.lcut(t['object'])
            for word in abstract:
                if word !=item['subject']:
                    temp.append(word)
                else:
                    temp.append(item['subject_id'])
            sentence.append(temp)
        else:
            temp.append(item['subject_id'])
            temp.append(t['predicate'])
            for w in jieba.lcut(t['object']):
                temp.append(w)
            sentence.append(temp)
train_tf = []
with open('data/raw_data/train.json','r') as f:
    for line in f:
        sentence.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))
        train_tf.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))
with open('data/raw_data/eval722.json','r') as f:
    for line in f:
        sentence.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))
        train_tf.append(jieba.lcut(eval(str(json.loads(line)).lower())['text']))

word_tf={}
for s in sentence:
    for w in s:
        if w not in word_tf:
            word_tf[w]=1
        else:
            word_tf[w]+=1

data = pd.read_pickle('data/step1.pkl').loc[:,['text_id','kb_id','label','m_id','train_mention']]
train_size = data.shape[0]

test_data = pd.read_pickle('data/step1_test.pkl').loc[:,['text_id','kb_id','m_id','train_mention']]

test_data['label'] = -1

data_all = data.append(test_data)

data_all.shape

# SVD降维后的TFIDF矩阵，n-gram 1-5

data_all['text'] = data_all['text_id'].apply(lambda x:train_data_dict[str(x)]['text'])

data_all.head()

group_f = data_all[['text_id','text']].drop_duplicates()
group_f['text_segment'] = group_f['text'].apply(lambda x:jieba.lcut(x))

dct = Dictionary(group_f['text_segment'])  # fit dictionary
corpus = [dct.doc2bow(line) for line in group_f['text_segment']]  # convert corpus to BoW format
model = TfidfModel(corpus)  # fit model

vector = model[corpus]
group_f['vector'] = vector

data_all = data_all.merge(group_f[['text_id','vector']],on='text_id')

def cal_tfidf(mention,vec):
    for v in vec:
        try:
            if dct.token2id[mention]==v[0]:
                return v[1]
        except:
            pass
            #print(mention)
            
    return -1
data_all['mention_tfidf'] = data_all.apply(lambda x:cal_tfidf(x.train_mention,x.vector),axis=1)
data_all['sum_tfidf'] = data_all['vector'].apply(lambda x:sum([i[1] for i in x]))

def cal_mention_tf(x):
    try:
        return word_tf[x]
    except:
        return -1

data_all['mention_tf'] = data_all['train_mention'].apply(lambda x:cal_mention_tf(x))

train_word_tf={}
for s in train_tf:
    for w in s:
        if w not in train_word_tf:
            train_word_tf[w]=1
        else:
            train_word_tf[w]+=1

def cal_train_mention_tf(x):
    try:
        return train_word_tf[x]
    except:
        return -1

data_all['train_mention_tf'] = data_all['train_mention'].apply(lambda x:cal_train_mention_tf(x))

data_group = data_all.groupby('train_mention')['label'].agg({'train_mention_count':'count'})
data_all = data_all.merge(data_group,on='train_mention',how='left')

data_all['ner_rate'] = data_all['train_mention_count']/data_all['train_mention_tf']

feature = ['mention_tfidf','sum_tfidf','mention_tf','train_mention_tf','train_mention_count','ner_rate']

train_data= data_all.iloc[0:train_size,:].reset_index(drop=True)
test_data= data_all.iloc[train_size:,:].reset_index(drop=True)

train_data.loc[:,feature].to_pickle('data/step3.pkl')
test_data.loc[:,feature].to_pickle('data/step3_test.pkl')

print(train_data.shape,test_data.shape)

# 根据kb提取topic，赋值给data_all #########################################

# lsi_model.save('data/lsi_model.pth')  # save model
#loaded_model = LsiModel.load(tmp_fname)  #

# # 应用于text
# dct = Dictionary(group_f['text_segment'])  # fit dictionary
# corpus = [dct.doc2bow(line) for line in group_f['text_segment']]

# 最大topic是否吻合，topic点积，最大topic，topic

# vectorized_corpus = model[corpus]

# topic_name = ['topics_%d'%i for i in range(50)]

# # vectorized = []
# for x in vectorized_corpus:
#     topic_arr = np.zeros(50)
#     for tuple_x in x:
#         #print(tuple_x)
#         topic_arr[tuple_x[0]] = tuple_x[1]
#     vectorized.append(topic_arr)

# vec_df = pd.DataFrame(vectorized,columns=topic_name)

# vec_df.head()

# vec_df['max_topic'] = vec_df.idxmax(axis=1).apply(lambda x:int(x.split('_')[-1]))
